refactor: move pager to another file
